REFACTOR: 좋아요 버튼 뷰로 빼기
