+++ Odya-iOS/Core/Community/View/PostContentView.swift
@@ -16,9 +16,6 @@ struct PostContentView: View {
   @StateObject private var followViewModel = FollowButtonViewModel()
   /// 팔로우 상태 토글
   @State private var followState: Bool
-  
-  // Like
-  @StateObject private var likeViewModel = CommunityLikeViewModel()
 
   // Feed Content
   let communityId: Int
@@ -70,17 +67,22 @@ struct PostContentView: View {
         .lineLimit(2)
 
       /// 장소, 좋아요, 댓글
-      VStack {
-        HStack {
-          locationView
-          Spacer()
-          likeView
-          commentView
-        }
-        .padding(8)
+      HStack {
+        locationView
+        Spacer()
+        CommunityLikeButton(
+          communityId: communityId,
+          likeState: isUserLiked,
+          likeCount: likeCount,
+          baseColor: Color.odya.label.assistive
+        )
+        .padding(.trailing, 12)
+        commentView
       }
+      .padding(8)
       .background(Color.odya.elevation.elev3)
       .cornerRadius(Radius.medium)
+      .frame(height: 40)
 
     }  // VStack
     .padding(.vertical, 16)
@@ -101,140 +103,6 @@ struct PostContentView: View {
         .detail2Style()
         .foregroundColor(Color.odya.label.assistive)
     }
-  }
-  
-  /// Like
-  
-  enum HeartImage: String {
-    case mOff = "heart-off-m"
-    case lOn = "heart-on-l"
-    case sOn = "heart-on-s"
-    case sOff = "heart-off-s"
-    
-    func nextOn() -> HeartImage {
-      switch self {
-      case .mOff: return .lOn
-      case .lOn: return .sOn
-      default: return .sOn
-      }
-    }
-    
-    func nextOff() -> HeartImage {
-      switch self {
-      case .sOn: return .sOff
-      case .sOff: return .mOff
-      default: return .mOff
-      }
-    }
-  }
-  
-  @State private var likeImageName = HeartImage.mOff
-  
-  private var likeView: some View {
-    HStack(spacing: 4) {
-      VStack {
-        if isUserLiked {
-          Image(likeImageName.rawValue)
-        } else {
-          Image(likeImageName.rawValue)
-            .renderingMode(.template)
-            .foregroundColor(Color.odya.label.assistive)
-  //            .padding(6)
-  //            .frame(width: 24, height: 24)
-        }
-      }
-      .onTapGesture {
-        print("좋아요 버튼 눌림!!!!")
-        if isUserLiked {
-          withAnimation(.spring()) {
-            // 좋아요 삭제
-            isUserLiked = false
-            likeCount -= 1
-            likeViewModel.deleteLike(communityId: communityId)
-            
-            likeImageName = likeImageName.nextOff()
-            print(likeImageName.rawValue)
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.4) {
-              likeImageName = likeImageName.nextOff()
-              print(likeImageName.rawValue)
-            }
-          }
-        } else {
-          withAnimation(.spring()) {
-            // 좋아요 생성
-            isUserLiked = true
-            likeCount += 1
-            likeViewModel.createLike(communityId: communityId)
-            
-            likeImageName = likeImageName.nextOn()
-            print(likeImageName.rawValue)
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.4) {
-              likeImageName = likeImageName.nextOn()
-              print(likeImageName.rawValue)
-            }
-          }
-        }
-      }
-      .onAppear {
-        likeImageName = isUserLiked ? HeartImage.sOn : HeartImage.mOff
-      }
-
-      
-      // 좋아요 버튼
-//      Button {
-//        if isUserLiked {
-//          withAnimation(.spring()) {
-//            // 좋아요 삭제
-//            isUserLiked = false
-//            likeCount -= 1
-//            likeViewModel.deleteLike(communityId: communityId)
-//
-//            likeImageName = likeImageName.nextOff()
-//            print(likeImageName.rawValue)
-//            DispatchQueue.main.asyncAfter(deadline: .now() + 0.4) {
-//              likeImageName = likeImageName.nextOff()
-//              print(likeImageName.rawValue)
-//            }
-//          }
-//        } else {
-//          withAnimation(.spring()) {
-//            // 좋아요 생성
-//            isUserLiked = true
-//            likeCount += 1
-//            likeViewModel.createLike(communityId: communityId)
-//
-//            likeImageName = likeImageName.nextOn()
-//            print(likeImageName.rawValue)
-//            DispatchQueue.main.asyncAfter(deadline: .now() + 0.4) {
-//              likeImageName = likeImageName.nextOn()
-//              print(likeImageName.rawValue)
-//            }
-//          }
-//        }
-//      } label: {
-//        if isUserLiked {
-//          Image(likeImageName.rawValue)
-//        } else {
-//          Image(likeImageName.rawValue)
-//            .renderingMode(.template)
-//            .foregroundColor(Color.odya.label.assistive)
-////            .padding(6)
-////            .frame(width: 24, height: 24)
-//        }
-//      }
-//      .onTapGesture {
-//        print("좋아요 버튼 눌림!!!!")
-//      }
-//      .onAppear {
-//        likeImageName = isUserLiked ? HeartImage.sOn : HeartImage.mOff
-//      }
-
-      // 좋아요 수
-      Text(likeCount > 99 ? "99+" : "\(likeCount)")
-        .detail1Style()
-        .foregroundColor(Color.odya.label.assistive)
-    }
-    .padding(.trailing, 12)
   }
 
   /// Comment
